rasa-chatbot/requirements.txt
rasa>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
dataclasses>=0.6
lxml>=4.9.0